+++ src/main/java/net/agadii/crystalinfused/block/CrystalPurifierBlock.java
@@ -81,7 +81,7 @@ public class CrystalPurifierBlock extends BlockWithEntity implements BlockEntity
 
     @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos,
-                              PlayerEntity player, Hand hand, BlockHitResult hit) {
+                              PlayerEntity player, BlockHitResult hit) {
         if (!world.isClient) {
             NamedScreenHandlerFactory screenHandlerFactory = ((CrystalPurifierBlockEntity) world.getBlockEntity(pos));