+++ src/main/java/net/agadii/crystalinfused/recipe/ModRecipes.java
@@ -7,14 +7,14 @@ import net.minecraft.util.Identifier;
 
 public class ModRecipes {
     public static void registerRecipes() {
-        Registry.register(Registries.RECIPE_SERIALIZER, new Identifier(CrystalInfused.MOD_ID, CrystalPurificationRecipe.Serializer.ID),
+        Registry.register(Registries.RECIPE_SERIALIZER, Identifier.of(CrystalInfused.MOD_ID, CrystalPurificationRecipe.Serializer.ID),
                 CrystalPurificationRecipe.Serializer.INSTANCE);
-        Registry.register(Registries.RECIPE_TYPE, new Identifier(CrystalInfused.MOD_ID, CrystalPurificationRecipe.Type.ID),
+        Registry.register(Registries.RECIPE_TYPE, Identifier.of(CrystalInfused.MOD_ID, CrystalPurificationRecipe.Type.ID),
                 CrystalPurificationRecipe.Type.INSTANCE);
 
-        Registry.register(Registries.RECIPE_SERIALIZER, new Identifier(CrystalInfused.MOD_ID, CrystalInfusionRecipe.Serializer.ID),
+        Registry.register(Registries.RECIPE_SERIALIZER, Identifier.of(CrystalInfused.MOD_ID, CrystalInfusionRecipe.Serializer.ID),
                 CrystalInfusionRecipe.Serializer.INSTANCE);
-        Registry.register(Registries.RECIPE_TYPE, new Identifier(CrystalInfused.MOD_ID, CrystalInfusionRecipe.Type.ID),
+        Registry.register(Registries.RECIPE_TYPE, Identifier.of(CrystalInfused.MOD_ID, CrystalInfusionRecipe.Type.ID),
                 CrystalInfusionRecipe.Type.INSTANCE);
     }
 }