+++ src/main/java/net/agadii/crystalinfused/screen/ModScreenHandlers.java
@@ -1,24 +1,33 @@
 package net.agadii.crystalinfused.screen;
 
 import net.agadii.crystalinfused.CrystalInfused;
+import net.agadii.crystalinfused.block.entity.CrystalPurifierBlockEntity;
 import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerType;
+import net.minecraft.network.PacketByteBuf;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.screen.ScreenHandlerType;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.math.BlockPos;
 
 public class ModScreenHandlers {
     public static ScreenHandlerType<CrystalPurificationScreenHandler> CRYSTAL_PURIFICATION_SCREEN_HANDLER =
-            new ExtendedScreenHandlerType<>(CrystalPurificationScreenHandler::new);
+            new ExtendedScreenHandlerType<>(
+                    CrystalPurificationScreenHandler::new,
+                    BlockPos.PACKET_CODEC
+            );
 
     public static ScreenHandlerType<CrystalInfusionScreenHandler> CRYSTAL_INFUSION_SCREEN_HANDLER =
-            new ExtendedScreenHandlerType<>(CrystalInfusionScreenHandler::new);
+            new ExtendedScreenHandlerType<>(
+                    CrystalInfusionScreenHandler::new,
+                    BlockPos.PACKET_CODEC
+            );
 
     public static void registerAllScreenHandlers() {
-        Registry.register(Registries.SCREEN_HANDLER, new Identifier(CrystalInfused.MOD_ID, "crystal_purifier"),
+        Registry.register(Registries.SCREEN_HANDLER, Identifier.of(CrystalInfused.MOD_ID, "crystal_purifier"),
                 CRYSTAL_PURIFICATION_SCREEN_HANDLER);
 
-        Registry.register(Registries.SCREEN_HANDLER, new Identifier(CrystalInfused.MOD_ID, "crystal_infuser"),
+        Registry.register(Registries.SCREEN_HANDLER, Identifier.of(CrystalInfused.MOD_ID, "crystal_infuser"),
                 CRYSTAL_INFUSION_SCREEN_HANDLER);
     }
 }