+++ src/main/java/utils/CodecUtils.java
@@ -2,57 +2,56 @@ package utils;
 
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.DataResult;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.type.ItemEnchantmentsComponent;
+import net.minecraft.component.type.NbtComponent;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.StringNbtReader;
 import net.minecraft.recipe.Ingredient;
 import net.minecraft.registry.Registries;
+import net.minecraft.util.dynamic.Codecs;
 
+import java.util.List;
 import java.util.Optional;
 
 public class CodecUtils {
-    public static final Codec<Ingredient> INGREDIENT_WITH_NBT_CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            Registries.ITEM.getCodec().fieldOf("item").forGetter(ing -> ing.getMatchingStacks()[0].getItem()),
-            Codec.STRING.optionalFieldOf("nbt").forGetter(ing -> {
-                ItemStack stack = ing.getMatchingStacks()[0];
-                NbtCompound nbt = stack.getNbt();
+    public static final Codec<Ingredient> INGREDIENT_FROM_STACK_CODEC = ItemStack.CODEC.flatComapMap(
+            Ingredient::ofStacks,
+            (Ingredient ingredient) -> {
+                ItemStack[] items = ingredient.getMatchingStacks();
 
-                return java.util.Optional.ofNullable(nbt != null ? nbt.toString() : null);
-            })
-    ).apply(instance, (item, nbtString) -> {
-        ItemStack stack = new ItemStack(item);
-        nbtString.ifPresent(s -> {
-            try {
-                stack.setNbt(StringNbtReader.parse(s));
-            } catch (CommandSyntaxException e) {
-                throw new RuntimeException("Invalid NBT: " + s, e);
+                if (items.length > 0) {
+                    return DataResult.success(items[0]);
+                }
+                return DataResult.error(() -> "Cannot encode empty ingredient");
             }
-        });
+    );
 
-        return Ingredient.ofStacks(stack);
-    }));
-
-    public static final Codec<ItemStack> RECIPE_RESULT_WITH_NBT_CODEC = RecordCodecBuilder.create(instance -> instance.group(
-            Registries.ITEM.getCodec().fieldOf("item").forGetter(ItemStack::getItem),
-            Codec.INT.optionalFieldOf("count", 1).forGetter(ItemStack::getCount),
-            Codec.STRING.optionalFieldOf("nbt").forGetter(stack -> {
-                NbtCompound nbt = stack.getNbt();
-
-                return Optional.ofNullable(nbt != null ? nbt.toString() : null);
-            })
-    ).apply(instance, (item, count, nbtString) -> {
-        ItemStack stack = new ItemStack(item, count);
-        nbtString.ifPresent(s -> {
-            try {
-                stack.setNbt(StringNbtReader.parse(s));
-            } catch (CommandSyntaxException e) {
-                throw new RuntimeException("Invalid NBT: " + s, e);
-            }
-        });
-
-        return stack;
-    }));
+    public static Codec<List<Ingredient>> validateAmount(Codec<Ingredient> delegate, int max) {
+        return delegate.listOf().flatXmap(
+                list -> {
+                    if (list.isEmpty()) {
+                        return DataResult.error(() -> "Recipe has no ingredients!");
+                    }
+                    if (list.size() > max) {
+                        return DataResult.error(() -> "Recipe has too many ingredients!");
+                    }
+                    return DataResult.success(list);
+                },
+                list -> {
+                    if (list.isEmpty()) {
+                        return DataResult.error(() -> "Recipe has no ingredients!");
+                    }
+                    if (list.size() > max) {
+                        return DataResult.error(() -> "Recipe has too many ingredients!");
+                    }
+                    return DataResult.success(list);
+                }
+        );
+    }
 }